requirements.txt
pdfplumber
pandas
requests
openpyxl
spacy
sentence-transformers
scikit-learn
aiohttp
python-dotenv
google-generativeai
psycopg2-binary
faiss-cpu